quick fix tests
